ilp64: make the Dis word 64-bit (Limbo int == pointer == WORD == 8)
Convert the LP64 (32-bit-int / 64-bit-pointer, decoupled) ABI to ILP64:
a Dis word, a Limbo int, a pointer/register slot, a real and a big are now
all 8 bytes, restoring Inferno's original `pointer == WORD` invariant at
64 bits.  This collapses the WORD!=PTR split that was the source of the
high-address heap-corruption bug class.  caerwynj/inferno64 used as the
reference for each puzzle; the conversion is small here because the tree
was already 64-bit-pointer-clean.

Core flip:
  * interp.h: WORD/UWORD -> long/unsigned long.
  * isa.h:    IBY2WD 4 -> 8 (now == IBY2PTR == IBY2LG).  XMAGIC8 tagging
              stays valid and self-consistent.

Fallout fixed (each caught by running the system + tests/lp64):
  * ecom.c: array-index opcode select — IINDW and IINDL collided at size 8;
            a word now indexes like a long (IINDL), IINDW is dead.  Matches
            caerwynj.
  * runt.c: Limbo print %d/%o/%x truncated 64-bit ints to 32 (printed
            4e9 as -294967296).  Promote the int verb to %l<verb> so the
            fmt library reads a full long.  (Latent in caerwynj too.)
  * load.c: DEFW data items (.dis int constants) loaded via disw() were
            zero-extended into the 64-bit WORD, so a negative int (-2)
            became 0x00000000FFFFFFFE = 4294967294 -> wild array indices /
            "array bounds error" (this is what broke sh).  Sign-extend the
            low 32 bits.
  * interp.h: struct Alt nsend/nrecv were C int (4B) but the compiler emits
            the alt header as two word-sized (tint, now 8B) counts, so the
            channel array ac[] landed at the wrong offset -> alt deadlock.
            Make them WORD (caerwynj uses intptr).

Validation (tests/lp64, interpreted): 177 ok / 1 FAIL / 0 err across 9
suites (was 178/178 on LP64 master).  vm 40/40, concur 11/11, crypto 18/18,
plumb 48/48, selfhost 9/9, loader/except/modglobal all green.  The one
failure (30_styxnet Tmsg.Attach) is the NOFID sentinel: `NOFID con int ~0`
is 64-bit all-ones, but the 32-bit wire fid round-trips to 0xFFFFFFFF.
Correct fix needs true-64-bit int constant folding (limbo.y threshold +
nodes.c masks), a deliberate tree-wide change deferred to its own pass;
the same definition is latent in caerwynj's tree.

JIT (-c1) not yet revalidated for ILP64; interpreter only.

+++ include/interp.h
@@ -1,6 +1,6 @@
 typedef uchar		BYTE;		/* 8  bits */
-typedef int		WORD;		/* 32 bits */
-typedef unsigned int	UWORD;		/* 32 bits */
+typedef long		WORD;		/* 64 bits (ILP64: Limbo int) */
+typedef unsigned long	UWORD;		/* 64 bits */
 typedef vlong		LONG;		/* 64 bits */
 typedef uvlong		ULONG;		/* 64 bits */
 typedef double		REAL;		/* 64 double IEEE754 */
@@ -191,8 +191,8 @@ struct Altc
 
 struct Alt
 {
-	int	nsend;
-	int	nrecv;
+	WORD	nsend;		/* ILP64: the compiler emits the alt header as two */
+	WORD	nrecv;		/* word-sized (tint) counts, so ac[] sits at 2*IBY2WD */
 	Altc	ac[1];
 };
